refactor(lowering): extract lowered asm stream helpers
